materials/requirements.txt
fastapi
pydantic
uvicorn
peewee
requests
pymongo
python-dotenv
grpcio
grpcio-tools
pyjwt
python-multipart
uuid
